refactor(typewriter): Modularize Typewriter.nvim plugin
- Split init.lua into multiple modular files for better organization and maintainability.
- Created config.lua for configuration management.
- Created utils.lua for utility functions.
- Created commands.lua for main plugin commands.
- Created autocommands.lua to set up autocommands.
- Integrated center_block_config.lua for significant code block configurations.
- Updated each file with appropriate comment descriptions for clarity.

This refactor improves the structure and readability of the code, making it easier to manage and extend.
